Move tests into test dir in repo root
