+++ test/commands/list.test.ts
