+++ tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig",
-  "exclude": ["**/tests/**", "*.test.*", "src/types/**"]
+  "exclude": ["src/types/**"]
 }
